Add the ability to select the protocol
Some cards are dual protocol T=0 and T=1. Maybe you want to select the
protocol to be used instead of letting PC/SC selecting the protocol.

It may be useful because PC/SC on GNU/Linux prefers T=1 over T=0. But
PC/SC on Mac OS X prefers T=0 over T=1.

A code that was working on GNU/Linux failed to work on Mac OS X because
T=0 was used instdead of T=1.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/HandlerTest@7002 0ce88b0d-b2fd-0310-8134-9614164e65ea